Fixes in variables and cosmetic changes
Changes to fix issues in bastion_bootstrap.sh with variables not set and some cosmetic changes for cleaner code and generated config files.

+++ scripts/bastion_bootstrap.sh
@@ -14,7 +14,7 @@ function checkos () {
     if [[ "${unamestr}" == 'Linux' ]]; then
         platform='linux'
     else
-        echo "[WARNING] This script is not supported on MacOS or freebsd"
+        echo "[WARNING] This script is not supported on MacOS or FreeBSD"
         exit 1
     fi
     echo "${FUNCNAME[0]} Ended"
@@ -44,7 +44,9 @@ function setup_environment_variables() {
   BASTION_LOGFILE="${BASTION_MNT}/${BASTION_LOG}"
   BASTION_LOGFILE_SHADOW="${BASTION_MNT}/.${BASTION_LOG}"
   touch ${BASTION_LOGFILE}
-  ln ${BASTION_LOGFILE} ${BASTION_LOGFILE_SHADOW}
+  if ! [ -L "$BASTION_LOGFILE_SHADOW" ]; then
+    ln ${BASTION_LOGFILE} ${BASTION_LOGFILE_SHADOW}
+  fi
   mkdir -p /usr/bin/bastion
   touch /tmp/messages
   chmod 770 /tmp/messages
@@ -72,7 +74,7 @@ function usage() {
 }
 
 function chkstatus () {
-    if [ $? -eq 0 ]
+    if [[ $? -eq 0 ]]
     then
         echo "Script [PASS]"
     else
@@ -83,11 +85,11 @@ function chkstatus () {
 
 function osrelease () {
     OS=`cat /etc/os-release | grep '^NAME=' |  tr -d \" | sed 's/\n//g' | sed 's/NAME=//g'`
-    if [ "${OS}" == "Ubuntu" ]; then
+    if [[ "${OS}" == "Ubuntu" ]]; then
         echo "Ubuntu"
-    elif [ "${OS}" == "Amazon Linux AMI" ] || [ "${OS}" == "Amazon Linux" ]; then
+    elif [[ "${OS}" == "Amazon Linux AMI" ]] || [[ "${OS}" == "Amazon Linux" ]]; then
         echo "AMZN"
-    elif [ "${OS}" == "CentOS Linux" ]; then
+    elif [[ "${OS}" == "CentOS Linux" ]]; then
         echo "CentOS"
     else
         echo "Operating System Not Found"
@@ -113,7 +115,7 @@ export Allow_SSH="ssh"
 export Allow_SCP="scp"
 if [[ -z $SSH_ORIGINAL_COMMAND ]] || [[ $SSH_ORIGINAL_COMMAND =~ ^$Allow_SSH ]] || [[ $SSH_ORIGINAL_COMMAND =~ ^$Allow_SCP ]]; then
 #Allow ssh to instance and log connection
-    if [ -z "$SSH_ORIGINAL_COMMAND" ]; then
+    if [[ -z "$SSH_ORIGINAL_COMMAND" ]]; then
         /bin/bash
         exit 0
     else
@@ -140,7 +142,7 @@ EOF
     chmod a+x /usr/bin/bastion/shell
 
     release=$(osrelease)
-    if [ "${release}" == "CentOS" ]; then
+    if [[ "${release}" == "CentOS" ]]; then
         semanage fcontext -a -t ssh_exec_t /usr/bin/bastion/shell
     fi
 
@@ -154,13 +156,13 @@ function amazon_os () {
     echo -e "\nDefaults env_keep += \"SSH_CLIENT\"" >>/etc/sudoers
 cat <<'EOF' >> /etc/bashrc
 #Added by linux bastion bootstrap
-declare -rx IP=$(echo $SSH_CLIENT | awk '{print $1}')
+declare -rx IP=\$(echo \$SSH_CLIENT | awk '{print \$1}')
 EOF
 
-    echo " declare -rx BASTION_LOG=${BASTION_MNT}/${BASTION_LOG}" >> /etc/bashrc
+    echo "declare -rx BASTION_LOG=${BASTION_LOGFILE}" >> /etc/bashrc
 
 cat <<'EOF' >> /etc/bashrc
-declare -rx PROMPT_COMMAND='history -a >(logger -t "ON: $(date)   [FROM]:${IP}   [USER]:${USER}   [PWD]:${PWD}" -s 2>>${BASTION_LOG})'
+declare -rx PROMPT_COMMAND='history -a >(logger -t "[ON]:\$(date)   [FROM]:\${IP}   [USER]:\${USER}   [PWD]:\${PWD}" -s 2>>\${BASTION_LOG})'
 EOF
     chown root:ec2-user  ${BASTION_MNT}
     chown root:ec2-user  ${BASTION_LOGFILE}
@@ -197,7 +199,7 @@ EOF
 
     #Restart awslogs service
     local OS=`cat /etc/os-release | grep '^NAME=' |  tr -d \" | sed 's/\n//g' | sed 's/NAME=//g'`
-    if [ "$OS"  == "Amazon Linux" ]; then # amazon linux 2
+    if [[ "$OS"  == "Amazon Linux" ]]; then # amazon linux 2
         systemctl start awslogsd.service
         systemctl enable awslogsd.service
     else
@@ -219,13 +221,13 @@ function ubuntu_os () {
     chown root:ubuntu /usr/bin/script
 cat <<'EOF' >> /etc/bash.bashrc
 #Added by linux bastion bootstrap
-declare -rx IP=$(who am i --ips|awk '{print $5}')
+declare -rx IP=\$(who am i --ips|awk '{print \$5}')
 EOF
 
-    echo " declare -rx BASTION_LOG=${BASTION_MNT}/${BASTION_LOG}" >> /etc/bash.bashrc
+    echo "declare -rx BASTION_LOG=${BASTION_LOGFILE}" >> /etc/bash.bashrc
 
 cat <<'EOF' >> /etc/bash.bashrc
-declare -rx PROMPT_COMMAND='history -a >(logger -t "ON: $(date)   [FROM]:${IP}   [USER]:${USER}   [PWD]:${PWD}" -s 2>>${BASTION_LOG})'
+declare -rx PROMPT_COMMAND='history -a >(logger -t "[ON]:\$(date)   [FROM]:\${IP}   [USER]:\${USER}   [PWD]:\${PWD}" -s 2>>\${BASTION_LOG})'
 EOF
     chown root:ubuntu ${BASTION_MNT}
     chown root:ubuntu  ${BASTION_LOGFILE}
@@ -258,7 +260,7 @@ EOF
 
     #Install Unit file for Ubuntu 16.04
     ubuntu=`cat /etc/os-release | grep VERSION_ID | tr -d \VERSION_ID=\"`
-    if [ "${ubuntu}" == "16.04" ]; then
+    if [[ "${ubuntu}" == "16.04" ]]; then
 cat <<'EOF' >> /etc/systemd/system/awslogs.service
 [Unit]
 Description=The CloudWatch Logs agent
@@ -270,7 +272,7 @@ Restart=always
 KillMode=process
 TimeoutSec=infinity
 PIDFile=/var/awslogs/state/awslogs.pid
-ExecStart=/var/awslogs/bin/awslogs-agent-launcher.sh --start --background --pidfile $PIDFILE --user awslogs --chuid awslogs &
+ExecStart=/var/awslogs/bin/awslogs-agent-launcher.sh --start --background --pidfile \$PIDFILE --user awslogs --chuid awslogs &
 
 [Install]
 WantedBy=multi-user.target
@@ -297,17 +299,17 @@ EOF
 
 function cent_os () {
     echo -e "\nDefaults env_keep += \"SSH_CLIENT\"" >>/etc/sudoers
-    echo -e "#Added by the Linux Bastion Bootstrap\ndeclare -rx IP=$(echo ${SSH_CLIENT} | awk '{print $1}')" >> /etc/bashrc
+    echo -e "#Added by the Linux Bastion Bootstrap\ndeclare -rx IP=\$(echo \${SSH_CLIENT} | awk '{print \$1}')" >> /etc/bashrc
 
-    echo "declare -rx BASTION_LOG=${BASTION_MNT}/${BASTION_LOG}" >> /etc/bashrc
+    echo "declare -rx BASTION_LOG=${BASTION_LOGFILE}" >> /etc/bashrc
 
     cat <<- EOF >> /etc/bashrc
-    declare -rx PROMPT_COMMAND='history -a >(logger -t "ON: $(date)   [FROM]:${IP}   [USER]:${USER}   [PWD]:${PWD}" -s 2>>${BASTION_LOG})'
+declare -rx PROMPT_COMMAND='history -a >(logger -t "[ON]:\$(date)   [FROM]:\${IP}   [USER]:\${USER}   [PWD]:\${PWD}" -s 2>>\${BASTION_LOG})'
 EOF
 
     chown root:centos ${BASTION_MNT}
     chown root:centos /usr/bin/script
-    chown root:centos  /var/log/bastion/bastion.log
+    chown root:centos /var/log/bastion/bastion.log
     chmod 770 /var/log/bastion/bastion.log
     touch /tmp/messages
     chown root:centos /tmp/messages
@@ -316,43 +318,43 @@ EOF
 
     # Install CloudWatch Log service on Centos Linux
     centos=`cat /etc/os-release | grep VERSION_ID | tr -d \VERSION_ID=\"`
-    if [ "${centos}" == "7" ]; then
-        echo "file = ${BASTION_LOGFILE_SHADOW}" >> /tmp/groupname.txt
-        echo "log_group_name = ${CWG}" >> /tmp/groupname.txt
+    if [[ "${centos}" == "7" ]]; then
 
         cat <<EOF >> ~/cloudwatchlog.conf
-        [general]
-        state_file = /var/awslogs/state/agent-state
-        use_gzip_http_content_encoding = true
-        logging_config_file = /var/awslogs/etc/awslogs.conf
+[general]
+state_file = /var/awslogs/state/agent-state
+use_gzip_http_content_encoding = true
+logging_config_file = /var/awslogs/etc/awslogs.conf
 
-        [/var/log/bastion]
-        datetime_format = %Y-%m-%d %H:%M:%S
-        file = /var/log/messages
-        buffer_duration = 5000
-        log_stream_name = {instance_id}
-        initial_position = start_of_file
+[/var/log/bastion]
+datetime_format = %Y-%m-%d %H:%M:%S
+time_zone = UTC
+file = ${BASTION_LOGFILE_SHADOW}
+buffer_duration = 5000
+log_group_name = ${CWG}
+log_stream_name = {instance_id}
+initial_position = start_of_file
 EOF
         cat /tmp/groupname.txt >> ~/cloudwatchlog.conf
 
         curl https://s3.amazonaws.com/aws-cloudwatch/downloads/latest/awslogs-agent-setup.py -O
         chmod +x ./awslogs-agent-setup.py
-        ./awslogs-agent-setup.py -n -r ${REGION} -c ~/cloudwatchlog.conf
+        ./awslogs-agent-setup.py -n -r ${REGION} -c ~/cloudwatchlog.conf --no-proxy 169.254.169.254
         cat << EOF >> /etc/systemd/system/awslogs.service
-        [Unit]
-        Description=The CloudWatch Logs agent
-        After=rc-local.service
+[Unit]
+Description=The CloudWatch Logs agent
+After=rc-local.service
 
-        [Service]
-        Type=simple
-        Restart=always
-        KillMode=process
-        TimeoutSec=infinity
-        PIDFile=/var/awslogs/state/awslogs.pid
-        ExecStart=/var/awslogs/bin/awslogs-agent-launcher.sh --start --background --pidfile $PIDFILE --user awslogs --chuid awslogs &
+[Service]
+Type=simple
+Restart=always
+KillMode=process
+TimeoutSec=infinity
+PIDFile=/var/awslogs/state/awslogs.pid
+ExecStart=/var/awslogs/bin/awslogs-agent-launcher.sh --start --background --pidfile \$PIDFILE --user awslogs --chuid awslogs &
 
-        [Install]
-        WantedBy=multi-user.target
+[Install]
+WantedBy=multi-user.target
 EOF
         service awslogs restart
         chkconfig awslogs on
@@ -403,7 +405,7 @@ function request_eip() {
     _determine_eip_assc_status ${PUBLIC_IP_ADDRESS}
     set -e
 
-    if [[ ${_eip_associated} -ne 1 ]]; then
+    if [[ ${_eip_associated} -eq 0 ]]; then
       echo "The Public IP address associated with eth0 (${PUBLIC_IP_ADDRESS}) is already an Elastic IP. Not proceeding further."
       exit 1
     fi
@@ -413,7 +415,7 @@ function request_eip() {
 
     for eip in "${EIP_ARRAY[@]}"; do
 
-      if [ "${eip}" == "Null" ]; then
+      if [[ "${eip}" == "Null" ]]; then
         echo "Detected a NULL Value, moving on."
         continue
       fi
@@ -425,7 +427,7 @@ function request_eip() {
       if [[ ${_eip_associated} -eq 0 ]]; then
         echo "Elastic IP [${eip}] already has an association. Moving on."
         let _eip_assigned_count+=1
-        if [ "${_eip_assigned_count}" -eq "${#EIP_ARRAY[@]}" ]; then
+        if [[ "${_eip_assigned_count}" -eq "${#EIP_ARRAY[@]}" ]]; then
           echo "All of the stack EIPs have been assigned (${_eip_assigned_count}/${#EIP_ARRAY[@]}). I can't assign anything else. Exiting."
           exit 1
         fi
@@ -441,7 +443,7 @@ function request_eip() {
       rc=$?
       set -e
 
-      if [ ${rc} -ne 0 ]; then
+      if [[ ${rc} -ne 0 ]]; then
 
         let _eip_assigned_count+=1
         continue
@@ -455,7 +457,7 @@ function request_eip() {
 
 function _query_assigned_public_ip() {
   # Note: ETH0 Only.
-  # - Does not distinquish between EIP and Standard IP. Need to cross-ref later.
+  # - Does not distinguish between EIP and Standard IP. Need to cross-ref later.
   echo "Querying the assigned public IP"
   PUBLIC_IP_ADDRESS=$(curl -sq 169.254.169.254/latest/meta-data/public-ipv4/${ETH0_MAC}/public-ipv4s/)
 }
@@ -481,7 +483,7 @@ function _determine_eip_assc_status(){
 function _determine_eip_allocation(){
   echo "Determining EIP Allocation for [${1}]"
   resource_id_length=$(aws ec2 describe-addresses --public-ips ${1} --output text --region ${REGION} | head -n 1 | awk {'print $2'} | sed 's/.*eipalloc-//')
-  if [ "${#resource_id_length}" -eq 17 ]; then
+  if [[ "${#resource_id_length}" -eq 17 ]]; then
       eip_allocation=$(aws ec2 describe-addresses --public-ips ${1} --output text --region ${REGION}| egrep 'eipalloc-([a-z0-9]{17})' -o)
   else
       eip_allocation=$(aws ec2 describe-addresses --public-ips ${1} --output text --region ${REGION}| egrep 'eipalloc-([a-z0-9]{8})' -o)
@@ -509,11 +511,11 @@ setup_environment_variables
 SSH_BANNER="LINUX BASTION"
 
 # Read the options from cli input
-TEMP=`getopt -o h:  --long help,banner:,enable:,tcp-forwarding:,x11-forwarding: -n $0 -- "$@"`
+TEMP=`getopt -o h --longoptions help,banner:,enable:,tcp-forwarding:,x11-forwarding: -n $0 -- "$@"`
 eval set -- "${TEMP}"
 
 
-if [ $# == 1 ] ; then echo "No input provided! type ($0 --help) to see usage help" >&2 ; exit 1 ; fi
+if [[ $# == 1 ]] ; then echo "No input provided! type ($0 --help) to see usage help" >&2 ; exit 1 ; fi
 
 # extract options and their arguments into variables.
 while true; do
@@ -550,14 +552,14 @@ done
 # BANNER CONFIGURATION
 BANNER_FILE="/etc/ssh_banner"
 if [[ ${ENABLE} == "true" ]];then
-    if [ -z ${BANNER_PATH} ];then
+    if [[ -z ${BANNER_PATH} ]];then
         echo "BANNER_PATH is null skipping ..."
     else
         echo "BANNER_PATH = ${BANNER_PATH}"
         echo "Creating Banner in ${BANNER_FILE}"
         echo "curl  -s ${BANNER_PATH} > ${BANNER_FILE}"
         curl  -s ${BANNER_PATH} > ${BANNER_FILE}
-        if [ ${BANNER_FILE} ] ;then
+        if [[ -e ${BANNER_FILE} ]] ;then
             echo "[INFO] Installing banner ... "
             echo -e "\n Banner ${BANNER_FILE}" >>/etc/ssh/sshd_config
         else
@@ -590,15 +592,15 @@ fi
 
 release=$(osrelease)
 # Ubuntu Linux
-if [ "${release}" == "Ubuntu" ]; then
+if [[ "${release}" == "Ubuntu" ]]; then
     #Call function for Ubuntu
     ubuntu_os
 # AMZN Linux
-elif [ "${release}" == "AMZN" ]; then
+elif [[ "${release}" == "AMZN" ]]; then
     #Call function for AMZN
     amazon_os
 # CentOS Linux
-elif [ "${release}" == "CentOS" ]; then
+elif [[ "${release}" == "CentOS" ]]; then
     #Call function for CentOS
     cent_os
 else
